Rewrite coupling tests
Simplified 3d test to have only 1 Hex8 element and twisting moment
in z-direction. This is way more simpler test, without having any
dependencies to AbaqusReader, thus it's not that fragile to break.

Also, rewrite tests so that they don't need JuliaFEM. Both 2d test and
3d test did have dependency to JuliaFEM because they used Elasticity
problem, this is now handled by saving the assembled stiffness matrices
to file which is then read to get results.